Better DataFrame#to_html for MultiIndex order plus factoring out some specs/lib code
