+++ lib/daru.rb
@@ -78,6 +78,7 @@ require 'daru/vector.rb'
 require 'daru/dataframe.rb'
 require 'daru/monkeys.rb'
 require 'daru/formatters/table'
+require 'daru/iruby/helpers'
 
 require 'daru/core/group_by.rb'
 require 'daru/core/query.rb'
