+++ lib/daru/iruby/templates/dataframe_mi.html.erb
@@ -2,14 +2,27 @@
   <tr>
     <th colspan='<%= @vectors.size+index.width %>'>Daru::DataFrame<%= name ? ": #{name} " : ''%>(<%=nrows%>x<%=ncols%>)</th>
   </tr>
-  <tr>
-    <th colspan="<%= index.width %>"></th>
-    <% @vectors.each do |vector| %>
-      <th><%=vector%></th>
+  <% if @vectors.is_a? MultiIndex %>
+    <% Daru::IRuby::Helpers.tuples_with_colspans(@vectors).each_with_index do |tuple, idx| %>
+      <tr>
+        <% if idx.zero? %>
+          <th colspan="<%= index.width %>" rowspan="<%= @vectors.width %>"></th>
+        <% end %>
+        <% tuple.each do |idx, span| %>
+          <th colspan="<%= span %>"><%= idx %></th>
+        <% end %>
+      </tr>
     <% end %>
-  </tr>
+  <% else %>
+    <tr>
+      <th colspan="<%= index.width %>"></th>
+      <% @vectors.each do |vector| %>
+        <th><%=vector%></th>
+      <% end %>
+    </tr>
+  <% end %>
 
-  <% @index.tuples_with_rowspans.first(threshold).zip(@index.to_a).each do |tuple, index| %>
+  <% Daru::IRuby::Helpers.tuples_with_rowspans(@index).first(threshold).zip(@index.to_a).each do |tuple, index| %>
     <tr>
       <% tuple.each do |idx, span| %>
         <th rowspan="<%= span %>"><%= idx %></th>
